--- appservice/.mdb/meta.json
@@ -1,7 +1,0 @@
-{
-    "config_version": 20210101,
-    "app_id": "6618e97c17e704236e8ad9b9",
-    "group_id": "6618e0ccdf1d53317bc45437",
-    "client_app_id": "application-1-rgjfz",
-    "last_pulled": 1712908676
-}
