--- appservice/auth/custom_user_data.json
@@ -1,3 +1,0 @@
-{
-    "enabled": false
-}
