--- appservice/auth/providers.json
@@ -1,7 +1,0 @@
-{
-    "api-key": {
-        "name": "api-key",
-        "type": "api-key",
-        "disabled": true
-    }
-}
