--- appservice/data_sources/mongodb-atlas/config.json
@@ -1,10 +1,0 @@
-{
-    "name": "mongodb-atlas",
-    "type": "mongodb-atlas",
-    "config": {
-        "clusterName": "whatscooking",
-        "readPreference": "primary",
-        "wireProtocolEnabled": false
-    },
-    "version": 1
-}
