--- appservice/environments/no-environment.json
@@ -1,3 +1,0 @@
-{
-    "values": {}
-}
