--- appservice/environments/production.json
@@ -1,3 +1,0 @@
-{
-    "values": {}
-}
