--- appservice/environments/testing.json
@@ -1,3 +1,0 @@
-{
-    "values": {}
-}
