--- appservice/functions/config.json
@@ -1,26 +1,0 @@
-[
-    {
-        "name": "getRestaurantsAutocomplete",
-        "private": false,
-        "run_as_system": true,
-        "disable_arg_logs": true
-    },
-    {
-        "name": "getRestaurants",
-        "private": false,
-        "run_as_system": true,
-        "disable_arg_logs": true
-    },
-    {
-        "name": "getFacets",
-        "private": false,
-        "run_as_system": true,
-        "disable_arg_logs": true
-    },
-    {
-        "name": "getFoodSynonyms",
-        "private": false,
-        "run_as_system": true,
-        "disable_arg_logs": true
-    }
-]
