--- appservice/graphql/config.json
@@ -1,4 +1,0 @@
-{
-    "use_natural_pluralization": true,
-    "disable_schema_introspection": false
-}
