--- appservice/graphql/validation_settings.json
@@ -1,6 +1,0 @@
-{
-    "read_validation_action": "ERROR",
-    "read_validation_level": "STRICT",
-    "write_validation_action": "ERROR",
-    "write_validation_level": "STRICT"
-}
