--- appservice/realm_config.json
@@ -1,7 +1,0 @@
-{
-    "app_id": "application-1-rgjfz",
-    "config_version": 20210101,
-    "name": "Application-1",
-    "provider_region": "aws-ap-southeast-1",
-    "deployment_model": "LOCAL"
-}
