--- appservice/sync/config.json
@@ -1,3 +1,0 @@
-{
-    "development_mode_enabled": false
-}
