--- appservice/values/ConnectionString.json
@@ -1,5 +1,0 @@
-{
-    "name": "ConnectionString",
-    "value": "connectionString",
-    "from_secret": true
-}
