+++ lambda/provider.tf
@@ -1,0 +1,4 @@
+# Here you can also customize your credentials settings, but i recommend to use the default file ~/.aws/credentials with different profiles.
+provider "aws" {
+  region = "ap-southeast-1"
+}
