+++ lambda/variables.tf
@@ -1,0 +1,4 @@
+variable "mongodb_uri" {
+  description = "Whatscooking MongoDB Atlas URI"
+  type        = string
+}
